requirements.txt
Flask
SQLAlchemy
Flask-WTF
Jinja2
better_profanity
Flask-Session
flask-marshmallow
